requirements.txt
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.75.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.14
pytest==8.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.6
starlette==0.17.1
typing_extensions==4.10.0
uvicorn==0.17.6
